+++ examples/account/evmapp/src/main/java/io/horizen/examples/EvmAppModule.java
@@ -30,13 +30,15 @@ public class EvmAppModule extends AccountAppModule {
     // 1 or 2 should be enough to avoid SC block reverting in the most cases.
     // WARNING. It must be constant and should not be changed inside Sidechain network
     private final int mcBlockRefDelay;
-    private boolean allForksEnabled;
+    private final boolean allForksEnabled;
+    private final int maxHistRewLen;
 
 
-    public EvmAppModule(String userSettingsFileName, int mcBlockDelayReference, boolean allForksEnabled) {
+    public EvmAppModule(String userSettingsFileName, int mcBlockDelayReference, boolean allForksEnabled, int maxHistRewLen) {
         this.settingsReader = new SettingsReader(userSettingsFileName, Optional.empty());
         this.mcBlockRefDelay = mcBlockDelayReference;
         this.allForksEnabled = allForksEnabled;
+        this.maxHistRewLen = maxHistRewLen;
     }
 
     @Override
@@ -115,5 +117,8 @@ public class EvmAppModule extends AccountAppModule {
         bind(Integer.class)
                 .annotatedWith(Names.named("MainchainBlockReferenceDelay"))
                 .toInstance(mcBlockRefDelay);
+        bind(Integer.class)
+                .annotatedWith(Names.named("MaxHistoryRewriteLength"))
+                .toInstance(maxHistRewLen);
     }
 }