+++ sdk/src/main/scala/io/horizen/account/state/ForgerStakeMsgProcessor.scala
@@ -247,6 +247,7 @@ case class ForgerStakeMsgProcessor(params: NetworkParams) extends NativeSmartCon
 
     var nodeReference = view.getAccountStorage(contractAddress, LinkedListTipKey)
 
+    var numOfMigratedElem = 0
     while (!linkedListNodeRefIsNull(nodeReference)) {
       val (item: AccountForgingStakeInfo, prevNodeReference: Array[Byte]) = getStakeListItem(view, nodeReference)
 
@@ -256,6 +257,7 @@ case class ForgerStakeMsgProcessor(params: NetworkParams) extends NativeSmartCon
 
       removeNode(view, item.stakeId, contractAddress)
       nodeReference = prevNodeReference
+      numOfMigratedElem = numOfMigratedElem + 1
     }
     view.removeAccountStorage(contractAddress, LinkedListTipKey)
 
@@ -263,7 +265,10 @@ case class ForgerStakeMsgProcessor(params: NetworkParams) extends NativeSmartCon
     val evmLog = getEthereumConsensusDataLog(upgradeEvent)
     view.addLog(evmLog)
 
+    log.info(s"Forger stakes storage upgraded successfully to version 2 - $numOfMigratedElem items migrated")
+
     ForgerStakeStorage.saveStorageVersion(view, ForgerStakeStorageVersion.VERSION_2)
+
 
   }