+++ sdk/src/main/scala/io/horizen/params/RegTestParams.scala
@@ -1,17 +1,17 @@
 package io.horizen.params
 
-import io.horizen.block.SidechainCreationVersions.{SidechainCreationVersion, SidechainCreationVersion1}
-
-import java.math.BigInteger
 import com.horizen.commitmenttreenative.CustomBitvectorElementsConfig
+import io.horizen.account.proposition.AddressProposition
+import io.horizen.block.SidechainCreationVersions.{SidechainCreationVersion, SidechainCreationVersion1}
 import io.horizen.cryptolibprovider.CircuitTypes
 import io.horizen.cryptolibprovider.CircuitTypes.CircuitTypes
 import io.horizen.cryptolibprovider.utils.CumulativeHashFunctions
+import io.horizen.history.AbstractHistory.MAX_HISTORY_REWRITING_LENGTH
 import io.horizen.proposition.{PublicKey25519Proposition, SchnorrProposition, VrfPublicKey}
 import sparkz.core.block.Block
-import sparkz.util.ModifierId
-import sparkz.util.bytesToId
-import scala.concurrent.duration._
+import sparkz.util.{ModifierId, bytesToId}
+
+import java.math.BigInteger
 
 case class RegTestParams(
                           override val sidechainId: Array[Byte] = new Array[Byte](32),
@@ -42,6 +42,8 @@ case class RegTestParams(
                           override val isHandlingTransactionsEnabled: Boolean = true,
                           override val mcBlockRefDelay: Int = 0,
                           override val resetModifiersStatus: Boolean = false,
+                          override val maxHistoryRewritingLength: Int = MAX_HISTORY_REWRITING_LENGTH,
+                          override val rewardAddress: Option[AddressProposition] = None,
                         ) extends NetworkParams {
   override val EquihashN: Int = 48
   override val EquihashK: Int = 5