+++ sdk/src/main/scala/io/horizen/utxo/SidechainApp.scala
@@ -68,7 +68,8 @@ class SidechainApp @Inject()
    @Named("ApplicationStopper") override val applicationStopper: SidechainAppStopper,
    @Named("ForkConfiguration") override val forkConfigurator: ForkConfigurator,
    @Named("AppVersion") appVersion: String,
-   @Named("MainchainBlockReferenceDelay") mcBlockReferenceDelay : Int
+   @Named("MainchainBlockReferenceDelay") mcBlockReferenceDelay : Int,
+   @Named("MaxHistoryRewriteLength") maxHistoryRewriteLength : Int
   )
   extends AbstractSidechainApp(
     sidechainSettings,
@@ -82,7 +83,8 @@ class SidechainApp @Inject()
       regtestId = 111,
       testnetId = 222,
       mainnetId = 333),
-    mcBlockReferenceDelay
+    mcBlockReferenceDelay,
+    maxHistoryRewriteLength
     )
 {