+++ sdk/src/test/java/io/horizen/account/transaction/EthereumTransactionRlpStreamCodecTest.java
@@ -5,9 +5,11 @@ import io.horizen.account.fixtures.EthereumTransactionFixture;
 import io.horizen.account.utils.EthereumTransactionDecoder;
 import io.horizen.account.utils.RlpStreamDecoder;
 import io.horizen.account.utils.RlpStreamEncoder;
-import io.horizen.transaction.exception.TransactionSemanticValidityException;
+import io.horizen.fork.ForkManagerUtil;
+import io.horizen.fork.SimpleForkConfigurator;
 import io.horizen.utils.BytesUtils;
 import org.bouncycastle.util.Arrays;
+import org.junit.Before;
 import org.junit.Test;
 import org.web3j.rlp.RlpList;
 import org.web3j.rlp.RlpString;
@@ -15,6 +17,7 @@ import sparkz.util.ByteArrayBuilder;
 import sparkz.util.serialization.Reader;
 import sparkz.util.serialization.VLQByteBufferReader;
 import sparkz.util.serialization.VLQByteBufferWriter;
+
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.util.stream.Stream;
@@ -23,6 +26,11 @@ import static io.horizen.account.utils.BigIntegerUInt256.getUnsignedByteArray;
 import static org.junit.Assert.*;
 
 public class EthereumTransactionRlpStreamCodecTest implements EthereumTransactionFixture {
+
+    @Before
+    public void setUp() {
+        ForkManagerUtil.initializeForkManager(new SimpleForkConfigurator(),"regtest");
+    }
 
     @Test
     public void rlpCodecEthTxLegacy() {
@@ -310,7 +318,6 @@ public class EthereumTransactionRlpStreamCodecTest implements EthereumTransactio
             Reader reader = new VLQByteBufferReader(ByteBuffer.wrap(b));
             try {
                 EthereumTransaction ethTx = EthereumTransactionDecoder.decode(reader);
-                //ethTx.semanticValidity();
                 if (reader.remaining() != 0)
                     throw new IllegalArgumentException("Stream is not consumed, we have " + reader.remaining() + "bytes left to read");
                 System.out.println(strings[0] + "--->" + ethTx.toString());
@@ -330,7 +337,7 @@ public class EthereumTransactionRlpStreamCodecTest implements EthereumTransactio
             Reader reader = new VLQByteBufferReader(ByteBuffer.wrap(b));
             try {
                 EthereumTransaction ethTx = EthereumTransactionDecoder.decode(reader);
-                ethTx.semanticValidity();
+                ethTx.semanticValidity(0);
                 System.out.println(strings[0] + "--->" + ethTx);
                 fail("Should not succeed");
 
@@ -362,7 +369,7 @@ public class EthereumTransactionRlpStreamCodecTest implements EthereumTransactio
         Reader reader = new VLQByteBufferReader(ByteBuffer.wrap(b));
         EthereumTransaction ethTx = EthereumTransactionDecoder.decode(reader);
         try {
-            ethTx.semanticValidity();
+            ethTx.semanticValidity(0);
         } catch (Exception e) {
             System.out.println(e.getMessage());
             fail("Should not fail");
@@ -381,7 +388,7 @@ public class EthereumTransactionRlpStreamCodecTest implements EthereumTransactio
         Reader reader = new VLQByteBufferReader(ByteBuffer.wrap(b));
         EthereumTransaction ethTx = EthereumTransactionDecoder.decode(reader);
         try {
-            ethTx.semanticValidity();
+            ethTx.semanticValidity(0);
         } catch (Exception e) {
             System.out.println(e.getMessage());
             fail("Should not fail");